+++ sdk/storage/azure_storage_queue/src/clients/queue_service_client.rs
@@ -32,6 +32,7 @@ impl QueueServiceClient {
         let mut options = options.unwrap_or_default();
         apply_storage_logging_defaults(&mut options.client_options);
 
+        let mut per_retry_policies: Vec<Arc<dyn Policy>> = Vec::default();
         if let Some(token_credential) = credential {
             if !endpoint.scheme().starts_with("https") {
                 return Err(azure_core::Error::with_message(
@@ -39,11 +40,10 @@ impl QueueServiceClient {
                     format!("{endpoint} must use https"),
                 ));
             }
-            let auth_policy: Arc<dyn Policy> = Arc::new(BearerTokenAuthorizationPolicy::new(
+            per_retry_policies.push(Arc::new(BearerTokenAuthorizationPolicy::new(
                 token_credential,
                 vec!["https://storage.azure.com/.default"],
-            ));
-            options.client_options.per_try_policies.push(auth_policy);
+            )));
         }
 
         let pipeline = Pipeline::new(
@@ -51,7 +51,7 @@ impl QueueServiceClient {
             option_env!("CARGO_PKG_VERSION"),
             options.client_options.clone(),
             Vec::default(),
-            Vec::default(),
+            per_retry_policies,
             None,
         );