+++ sdk/keyvault/azure_security_keyvault_keys/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- Vault URLs must use TLS (https).
+
 ### Bugs Fixed
 
 ### Other Changes